First submission of AnimalShelter package

+++ AnimalShelter/AdoptableAnimal.java
@@ -3,5 +3,5 @@ package AnimalShelter;
 public interface AdoptableAnimal {
     public String getName();
     public String getDescription();
-    public double getPrice();
+    public int getPrice();
 }
